2013-07-22  Alejandro Gallardo <alejandro.gallardo@artica.es>
	* include/functions_db.php: Changes on function
	"project_manager_check".
	
	* include/functions_form.php: Added an option in the
	function "combo_roles" to decide if show the manager.
	Added the function "combo_task_user_manager" wich shows
	a multiselect with the tasks with manage permission by
	the user.
	
	* include/functions_projects.php: New file. Added
	functions "get_projects_query", "get_tasks_query"
	and "get_project_access".
	
	* include/functions_tasks.php: Now the background of
	the tree is white.
	
	* include/help/en/help_project_roles.php
	  include/help/en/help_project_roles.php: New help files.
	  
	* include/js/jquery.multiselect.js: New file.
	JQuery widget for multiselect elements.
	
	* include/styles/flora/flora.multiselect.css: New file.
	Styles for the jquery.multiselect widget.
	
	* operation/contacts/contact_detail.php: Fixed a little
	error on the search query and on the acl.
	
	* operation/contracts/contract_detail.php: Fixed a
	little error on the search query.
	
	* operation/projects/gantt.php
	  operation/projects/milestones.php
	  operation/projects/people_manager.php
	  operation/projects/project.php
	  operation/projects/project_bubblegraph.php
	  operation/projects/project_detail.php
	  operation/projects/project_group_detail.php
	  operation/projects/project_tracking.php
	  operation/projects/project_tree.php
	  operation/projects/role_user_global.php
	  operation/projects/task.php
	  operation/projects/task_detail.php
	  operation/projects/task_workunit.php: Added or
	  updated the ACL. Minor fixes and changes on the
	  queries.

+++ operation/projects/milestones.php
@@ -13,28 +13,42 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
+// Load global vars
+global $config;
+
 if (check_login() != 0) {
 	audit_db("Noauth", $config["REMOTE_ADDR"], "No authenticated access", "Trying to access event viewer");
 	require ("general/noaccess.php");
 	exit;
 }
 
+include_once ("include/functions_projects.php");
+
+
 $id_user = $config["id_user"];
 $operation = get_parameter ("operation");
 $id_project = get_parameter ("id_project", -1);
+
+$project_access = get_project_access ($id_user, $id_project);
+
 if ($id_project != 1)
 	$project_name = get_db_value ("name", "tproject", "id", $id_project);
 else
 	$project_name = "";
 
-if ( $id_project == -1 ){
+if ($id_project == -1){
 	// Doesn't have access to this page
 	audit_db($id_user, $config["REMOTE_ADDR"], "ACL Violation","Trying to access to milestone manager without project");
 	no_permission();
 }
-
-if (! user_belong_project ($config["id_user"], $id_project)) {
+// ACL - To see the project milestones, you should have read access
+if (! $project_access["read"]) {
 	audit_db($id_user, $config["REMOTE_ADDR"], "ACL Violation","Trying to access to milestone manager without permissions");
+	no_permission();
+}
+// ACL - To manage a project milestone, you should have PM access
+if ($operation != "" && !$project_access["write"]) {
+	audit_db($id_user, $config["REMOTE_ADDR"], "ACL Violation","Trying to create or delete a milestone without permissions");
 	no_permission();
 }
 
@@ -115,7 +129,9 @@ if ($operation == ""){
 	echo "<th>".__('Milestone');
 	echo "<th>".__('Description');
 	echo "<th>".__('Timestamp');
-	echo "<th>".__('Delete');
+	if ($project_access['write']) {
+		echo "<th>".__('Delete');
+	}
 	$color=1;
 	$sql1="SELECT * FROM tmilestone WHERE id_project = $id_project";
 	if ($result=mysql_query($sql1))
@@ -140,15 +156,17 @@ if ($operation == ""){
 			echo $row["timestamp"];
 			
 			// DELETE
-			echo '<td class="'.$tdcolor.'" align="center">';
-			echo '<a href="index.php?sec=projects&sec2=operation/projects/milestones&id_project='.$id_project.'&operation=delete&id='.$row["id"].'" onClick="if (!confirm(\' '.__('Are you sure?').'\')) return false;"><img border=0 src="images/cross.png"></a>';
+			if ($project_access['write']) {
+				echo '<td class="'.$tdcolor.'" align="center">';
+				echo '<a href="index.php?sec=projects&sec2=operation/projects/milestones&id_project='.$id_project.'&operation=delete&id='.$row["id"].'" onClick="if (!confirm(\' '.__('Are you sure?').'\')) return false;"><img border=0 src="images/cross.png"></a>';
+			}
 			
 		}
 	echo "</table>";
 
     $project_manager = get_db_value ("id_owner", "tproject", "id", $id_project);
     // milestone creation
-    if ((give_acl($config["id_user"], 0, "PM")==1) OR ($project_manager == $config["id_user"])) {
+    if ($project_access['write']) {
 	    echo "<table class='button' width=90%>";
 	    echo "<tr><td align=right>";
